--- app/src/main/java/com/example/beerhive/beerlist/BeerContract.kt
@@ -1,16 +1,0 @@
-package com.example.beerhive.beerlist
-
-import com.example.beerhive.network.model.BeerResponse
-
-interface BeerContract {
-    interface BeerView {
-        fun displayBeerList(responseList: List<BeerResponse?>?)
-        fun showProgressDialog()
-        fun showErrorMessage()
-        fun dismissProgressDialog()
-    }
-
-    interface UserActionsListener {
-        fun loadBeerList()
-    }
-}
